backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
neo4j==5.17.0

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Redis / Celery (for async jobs)
redis==5.0.1
celery==5.3.6

# Validation & Utils
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
python-dateutil==2.8.2
python-dotenv==1.0.1

# Dev
pytest==8.0.1
pytest-asyncio==0.23.5
